proto: Flatten maps and sets into distinct Value variants
This change is mostly motivated by the difficulties that arise when
deserializing maps and sets, which could contain keys of different types. The
set of valid key types was already limited, so this is not functional change.

Additionally, this change simplifies the use of "naive" serializers based Serde,
which usually don't support complex keys (KeyValue). There are however still
other issues with this approach (like e.g. serde_json, which doesn't work with
boolean keys).