broker: Remove fifo size and use bounded channel for incoming messages
The internal logic used in both the broker and connection, which terminates
either when the fifo runs full, is far too rigid and triggers in perfectly valid
use cases, which just happen to cause short but large spikes in messages.

Also, the fact that the broker shuts down at all seems wrong now, because it's
always a client, which causes messages to pile up. All related logic should sit
in the connection.

The plan is thus to remove all related logic for now, and then later implement a
better strategy. This commit removes the broker fifo size and replaces the
channel for incoming messages with a bounded one.

Unfortunately, this is a rather intrusive change, because several BrokerHandle
methods now need to be async and take &mut self.