Move id types into aldrin-proto
This makes all id types, Object{Uuid,Cookie,Id} and Service{Uuid,Cookie,Id},
available centrally to all other crates.

The types are re-exported in aldrin-client for ease of use, because it is quite
common to only depend on that crate.

A new feature new-v4-ids is added to aldrin-proto, enabling the v4 uuid crate
feature and making new_v4 functions available on Object{Uuid,Cookie} and
Service{Uuid,Cookie}.

This is a breaking change, both on the API level as well as the protocol
level. The serialization format changed slightly because ServiceId now has a
different layout. The protocol version is bumped to 7.