proto: Extend handshake with a custom value exchange
Connect now allows the client to pass some custom data to the broker. The broker
can use this to accept or reject clients. To facilitate that, a Reject variant
is added to ConnnectReply. Both Ok and Reject carry Values.

Serialization tests have been extended accordingly.

The client and broker libraries do not yet handle this.