Remove aldrin-proto-macros crate
Serialize and Deserialize are intended to be implemented only via schemas and
codegen. Offering derive macros made it seem like using Aldrin without schemas
was an intended use-case.

This was further complicated by the fact, that macros were only offered for
struct and enum generation, but not for services.

In an attempt to consolidate use-cases towards schemas, these macros are
removed.